requirements.txt
asgiref==3.5.2
autopep8==2.0.1
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreschema==0.0.4
Django==4.0.8
django-cors-headers==3.13.0
django-filter==22.1
django-material-admin==1.8.6
django-split-settings==1.2.0
django-unfold==0.6.9
djangorestframework==3.14.0
drf-flex-fields==1.0.2
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==22.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.11
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.13
webencodings==0.5.1
zipp==3.16.0
